NativeTalk-master/backend/vps/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
openai-whisper==20231117
argostranslate==1.9.6
ffmpeg-python==0.2.0
